Added sublattice generators
It is often useful to find mappings from small qubit topologies into
larger ones.  We provide a set of generators, which enable a user to
easily adapt an emedding found in a small graph, to one found in a
larger graph.  This functionality can be used to implement a tiling
composite, for example.  We provide these functions to find Chimera
sublattices in Chimera, Pegasus and Zephyr, as well as Pegasus
sublattices in Pegasus and Zephyr sublattices in Zephyr.